chore(gRPC): remove duplicated header test code
