+++ crates/iota-e2e-tests/tests/grpc/v0/transaction_execution_service/header.rs
@@ -17,41 +17,8 @@ use iota_macros::sim_test;
 use iota_test_transaction_builder::make_transfer_iota_transaction;
 use iota_types::transaction::TransactionData;
 use test_cluster::TestClusterBuilder;
-use tonic::metadata::MetadataMap;
 
-/// Helper function to verify that all expected IOTA headers are present in the
-/// response metadata
-fn verify_iota_headers(metadata: &MetadataMap, operation_name: &str) {
-    let required_headers = [
-        headers::X_IOTA_CHAIN,
-        headers::X_IOTA_CHAIN_ID,
-        headers::X_IOTA_CHECKPOINT_HEIGHT,
-        headers::X_IOTA_EPOCH,
-        headers::X_IOTA_TIMESTAMP_MS,
-        headers::X_IOTA_TIMESTAMP,
-        headers::X_IOTA_LOWEST_AVAILABLE_CHECKPOINT,
-        headers::X_IOTA_LOWEST_AVAILABLE_CHECKPOINT_OBJECTS,
-        headers::X_IOTA_SERVER,
-    ];
-
-    for header_name in &required_headers {
-        assert!(
-            metadata.get(*header_name).is_some(),
-            "{operation_name} response should contain {header_name} header",
-        );
-    }
-}
-
-/// Helper function to parse a u64 header value
-fn parse_u64_header(metadata: &MetadataMap, header_name: &str) -> u64 {
-    metadata
-        .get(header_name)
-        .unwrap_or_else(|| panic!("{header_name} header should be present"))
-        .to_str()
-        .unwrap_or_else(|_| panic!("{header_name} header should be valid UTF-8"))
-        .parse()
-        .unwrap_or_else(|_| panic!("{header_name} header should be a valid u64"))
-}
+use crate::v0::header::{parse_u64_header, verify_iota_headers};
 
 #[sim_test]
 async fn test_response_headers() {
